+++ src/lib/db/migrations/0008_stormy_lenny_balinger.sql
@@ -1,3 +1,0 @@
-ALTER TABLE "user" ADD COLUMN "suspended_at" timestamp with time zone;--> statement-breakpoint
-ALTER TABLE "organization" ADD COLUMN "suspended_at" timestamp with time zone;--> statement-breakpoint
-ALTER TABLE "organization" ADD COLUMN "storage_quota_mb" integer;